ui: add base for user interface
It's just a draft for now and do nothing. But the main idea is here.
Inspired from 'Her' the film the main interface is composed by the
circle which represents RORI and moves.
Then the background to reflect the interaction. And then, messages
between RORI and the user will appears.